refactor(native): split chara_detail_scene_stitcher into .h/.cpp
First proof-of-concept of moving implementation out of the header-only native
backend, chosen as the lowest-risk target: no templates, no ONNX/WinRT deps.

- The header now carries declarations only. Frame is forward-declared (it is used
  solely by const reference in a private method), so consumers of this header no
  longer pull in the 429-line cv/frame.h.
- The new .cpp holds the moved bodies verbatim and includes its real
  dependencies explicitly (cv/frame.h, core/native_api.h, util/stds.h,
  logger_util.h, config). Previously these were satisfied only by the include
  order in native_api.cpp; the implementation is now self-contained.
- Both consuming CMake source lists are updated in lockstep: native/CMakeLists.txt
  (umacapture_cli) and windows/runner/CMakeLists.txt (the Flutter runner). Missing
  either would fail to link, so they must move together.

Verified: umacapture_cli builds, umacapture_tests still passes (12 cases / 22
assertions), `flutter build windows --debug` links the runner, and a stitch smoke
run on a copied fixture produced the expected output (onCharaDetailFinished
success:true).
