+++ native/CMakeLists.txt
@@ -18,6 +18,7 @@ set(
         SOURCE_FILES
         src/core/cli.cpp
         src/core/native_api.cpp
+        src/chara_detail/chara_detail_scene_stitcher.cpp
         src/condition/serializer.cpp
         src/util/logger_util.cpp
 )
