+++ windows/runner/CMakeLists.txt
@@ -24,6 +24,7 @@ set(
         "Runner.rc"
         "runner.exe.manifest"
         "${NATIVE_EXTENSION_DIR}/src/core/native_api.cpp"
+        "${NATIVE_EXTENSION_DIR}/src/chara_detail/chara_detail_scene_stitcher.cpp"
         "${NATIVE_EXTENSION_DIR}/src/condition/serializer.cpp"
         "${NATIVE_EXTENSION_DIR}/src/util/logger_util.cpp"
 )
